Added Reset Button to MapWidget.
From now on when the button is clicked all recent created Markers and Clusterers will be closed and the Map will zoom out and recenter to default position.

Some minor bug fixes. resize() function for markers now works as it should. -> 'When a marker is clicked. map centers and zooms to marker'